requirements.txt
Flask==2.3.2
Flask-PyMongo==2.3.0
Flask-JWT-Extended==4.5.2
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==0.21.0
pytz==2023.3
PyJWT==2.8.0
Werkzeug==2.3.7
Pillow==10.0.0
python-magic==0.4.27